Fix DEFAULT_SSID used before initialization
Problem: setup, line 310 - DEFAULT_SSID is uninitialized when first calling loadConfiguration.
As a result, when the XRP board is powered up for the first time, a config file is created with a null SSID.
This causes the board to fail to create an access point. This prevents the broken config file from being edited through a web interface.